Final polish: fix autoboxing, final fields, null-safe getters, lineSeparator
- BondChange: make bondChangeDelta final, remove setter, inline computation
- MCSSolution: null-safe getters (return 0/0.0 instead of null)
- AtomContainerComparatorBy2DCenter: Double.compare() instead of autoboxing
- MinSelection: fix Double.valueOf(int) == double comparison
- BondEnergies: diamond operator
- TestUtility: make DEBUG field final
- Replace System.getProperty("line.separator") with System.lineSeparator()
  across 7 files

All 135 tests pass.

Author: Syed Asad Rahman <asad.rahman@bioinceptionlabs.com>

+++ src/main/java/com/bioinceptionlabs/reactionblast/mapping/algorithm/checks/MinSelection.java
@@ -186,7 +186,7 @@ public class MinSelection extends Selector implements IResult {
     }
 
     private boolean isMCSSubgraph(IAtomContainer educt, double mcsSize) throws CDKException {
-        return Double.valueOf(educt.getAtomCount()) == mcsSize;
+        return educt.getAtomCount() == (int) mcsSize;
     }
 
     /**
